Refactoring js, splitting into multiple files
