+++ pretix_eth/templates/pretix_eth/web3modal.html
@@ -1,6 +1,6 @@
 {% csrf_token %}
 <script src="/static/3rd_party/web3modal/index.js"></script>
-<script type="module" src="/static/pretix_eth/web3modal.js"></script>
+<script type="module" src="/static/pretix_eth/web3modal/web3modal.js"></script>
 
 <div id="web3modal">
     <div id="prepare" class="pretix-eth-payment-steps">
